requirements.txt
Django==5.1.6
django_allauth==65.3.0
django_taggit==6.1.0
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
python-decouple==3.8
reportlab==4.3.1
